References

Public API

IntegrationTests.depending_projects — Function

depending_projects(
    package_name::AbstractString, 
    package_filter::Union{<:AbstractString,Regex}
    project_tree::AbstractDict=PkgDependency.builddict(Pkg.project().uuid, Pkg.project())
) -> Vector{String}

Returns a list of packages that have the package package_name as a dependency.

Arguments

package_name: Name of the dependency

package_filter: Ignore all packages that do not match package_filter. This includes the top node package of the graph. Child nodes are always checked for package_name, but they are not traversed if they do not match package_filter.

project_tree: Project tree in which to search for dependent packages. Each (sub-)package needs to be AbstractDict{String, AbstractDict}

Returns

A Vector{String} containing the names of all packages that have the given dependency.

IntegrationTests.depending_projects — Function

depending_projects(
    package_name::AbstractString, 
    package_filter::AbstractVector{<:AbstractString}, 
    project_tree::AbstractDict=PkgDependency.builddict(Pkg.project().uuid, Pkg.project())
) -> Vector{String}

Internal API

IntegrationTests._match_package_filter — Method

_match_package_filter(
    package_filter::AbstractVector{<:AbstractString}, 
    package::AbstractString
)::Bool

IntegrationTests._match_package_filter — Method

_match_package_filter(
    package_filter::Union{<:AbstractString,Regex}, 
    package::AbstractString
)::Bool

Check if package_filter contains package. Wrapper function for contains() and in().

Returns

true if it matches.

IntegrationTests._traverse_tree! — Method

_traverse_tree!(
    package_name::AbstractString, 
    package_filter::Union{<:AbstractString,Regex,AbstractVector{<:AbstractString}}, 
    project_tree::AbstractDict, 
    packages::AbstractVector{<:AbstractString},
    visited_packages::AbstractVector{<:AbstractString}
)

Traverse the project tree and add any package to packages that has the package package_name as a dependency.

Arguments

package_name: Name of the dependency

package_filter: Ignore all packages that do not match package_filter. This includes the top node package of the graph. Child nodes always are checked for package_name, but they are not traveres if they do not match package_filter.

project_tree: Project tree, where to search the dependent packages. Each (sub-) package needs to be AbstractDict{String, AbstractDict}

packages: Packages which has package_name as dependency.

visited_packages: List of packages that are not traveresed again. Avoids circular, infinite traversing.